sync: sync all payment updates
Rather than syncing payments based on the payment time, sync based on
the created and updated indices in core lightning. This ensures that all
updates are retrieved, also when a payment is updated after the fact. To
do this, listsendpays is called, rather than listpays. listsendpays
contains a filter based on index. Starting off with the last synced
index, fetch all missing newly created payments and updates. It's
possible that this particular sync doesn't fetch all payment parts
associated to a payment. Therefore, the sendpays are cached into a local
database. When a new sync round is done, all known existing parts are
fetched from the local database as well, to ensure the resulting sdk
payments are 'complete', i.e. they're not missing any parts. This should
fix the issue where for closed channels, payments would be marked
forever pending.